Move metadata to metadata.rs and use friendlier names
