+++ rust/pkg/cardano_serialization_lib.js.flow
@@ -6,6 +6,13 @@
  */
 
 /**
+ * @param {Transaction} tx
+ * @param {LinearFee} linear_fee
+ * @returns {BigNum}
+ */
+declare export function min_fee(tx: Transaction, linear_fee: LinearFee): BigNum;
+
+/**
  * @param {TransactionHash} tx_body_hash
  * @param {ByronAddress} addr
  * @param {LegacyDaedalusPrivateKey} key
@@ -80,13 +87,6 @@ declare export function get_deposit(
 ): BigNum;
 
 /**
- * @param {Transaction} tx
- * @param {LinearFee} linear_fee
- * @returns {BigNum}
- */
-declare export function min_fee(tx: Transaction, linear_fee: LinearFee): BigNum;
-
-/**
  */
 
 declare export var CertificateKind: {|
@@ -119,22 +119,22 @@ declare export var RelayKind: {|
 /**
  */
 
-declare export var TransactionMetadatumKind: {|
-  +MapTransactionMetadatumToTransactionMetadatum: 0, // 0
-  +ArrTransactionMetadatum: 1, // 1
-  +Int: 2, // 2
-  +Bytes: 3, // 3
-  +Text: 4, // 4
-|};
-
-/**
- */
-
 declare export var MultisigScriptKind: {|
   +MsigPubkey: 0, // 0
   +MsigAll: 1, // 1
   +MsigAny: 2, // 2
   +MsigNOfK: 3, // 3
+|};
+
+/**
+ */
+
+declare export var TransactionMetadatumKind: {|
+  +MetadataMap: 0, // 0
+  +MetadataList: 1, // 1
+  +Int: 2, // 2
+  +Bytes: 3, // 3
+  +Text: 4, // 4
 |};
 
 /**
@@ -1154,6 +1154,21 @@ declare export class Int {
    * @returns {Int}
    */
   static new_negative(x: BigNum): Int;
+
+  /**
+   * @returns {boolean}
+   */
+  is_positive(): boolean;
+
+  /**
+   * @returns {BigNum | void}
+   */
+  as_positive(): BigNum | void;
+
+  /**
+   * @returns {BigNum | void}
+   */
+  as_negative(): BigNum | void;
 }
 /**
  */
@@ -1309,7 +1324,7 @@ declare export class MapTransactionIndexToTransactionMetadata {
 }
 /**
  */
-declare export class MapTransactionMetadatumToTransactionMetadatum {
+declare export class MetadataHash {
   free(): void;
 
   /**
@@ -1319,16 +1334,67 @@ declare export class MapTransactionMetadatumToTransactionMetadatum {
 
   /**
    * @param {Uint8Array} bytes
-   * @returns {MapTransactionMetadatumToTransactionMetadatum}
+   * @returns {MetadataHash}
    */
-  static from_bytes(
-    bytes: Uint8Array
-  ): MapTransactionMetadatumToTransactionMetadatum;
+  static from_bytes(bytes: Uint8Array): MetadataHash;
+}
+/**
+ */
+declare export class MetadataList {
+  free(): void;
 
   /**
-   * @returns {MapTransactionMetadatumToTransactionMetadatum}
+   * @returns {Uint8Array}
    */
-  static new(): MapTransactionMetadatumToTransactionMetadatum;
+  to_bytes(): Uint8Array;
+
+  /**
+   * @param {Uint8Array} bytes
+   * @returns {MetadataList}
+   */
+  static from_bytes(bytes: Uint8Array): MetadataList;
+
+  /**
+   * @returns {MetadataList}
+   */
+  static new(): MetadataList;
+
+  /**
+   * @returns {number}
+   */
+  len(): number;
+
+  /**
+   * @param {number} index
+   * @returns {TransactionMetadatum}
+   */
+  get(index: number): TransactionMetadatum;
+
+  /**
+   * @param {TransactionMetadatum} elem
+   */
+  add(elem: TransactionMetadatum): void;
+}
+/**
+ */
+declare export class MetadataMap {
+  free(): void;
+
+  /**
+   * @returns {Uint8Array}
+   */
+  to_bytes(): Uint8Array;
+
+  /**
+   * @param {Uint8Array} bytes
+   * @returns {MetadataMap}
+   */
+  static from_bytes(bytes: Uint8Array): MetadataMap;
+
+  /**
+   * @returns {MetadataMap}
+   */
+  static new(): MetadataMap;
 
   /**
    * @returns {number}
@@ -1352,25 +1418,9 @@ declare export class MapTransactionMetadatumToTransactionMetadatum {
   get(key: TransactionMetadatum): TransactionMetadatum | void;
 
   /**
-   * @returns {TransactionMetadatums}
+   * @returns {MetadataList}
    */
-  keys(): TransactionMetadatums;
-}
-/**
- */
-declare export class MetadataHash {
-  free(): void;
-
-  /**
-   * @returns {Uint8Array}
-   */
-  to_bytes(): Uint8Array;
-
-  /**
-   * @param {Uint8Array} bytes
-   * @returns {MetadataHash}
-   */
-  static from_bytes(bytes: Uint8Array): MetadataHash;
+  keys(): MetadataList;
 }
 /**
  */
@@ -3303,20 +3353,16 @@ declare export class TransactionMetadatum {
   static from_bytes(bytes: Uint8Array): TransactionMetadatum;
 
   /**
-   * @param {MapTransactionMetadatumToTransactionMetadatum} map_transaction_metadatum_to_transaction_metadatum
+   * @param {MetadataMap} map
    * @returns {TransactionMetadatum}
    */
-  static new_map_transaction_metadatum_to_transaction_metadatum(
-    map_transaction_metadatum_to_transaction_metadatum: MapTransactionMetadatumToTransactionMetadatum
-  ): TransactionMetadatum;
+  static new_map(map: MetadataMap): TransactionMetadatum;
 
   /**
-   * @param {TransactionMetadatums} arr_transaction_metadatum
+   * @param {MetadataList} list
    * @returns {TransactionMetadatum}
    */
-  static new_arr_transaction_metadatum(
-    arr_transaction_metadatum: TransactionMetadatums
-  ): TransactionMetadatum;
+  static new_list(list: MetadataList): TransactionMetadatum;
 
   /**
    * @param {Int} int
@@ -3342,14 +3388,14 @@ declare export class TransactionMetadatum {
   kind(): number;
 
   /**
-   * @returns {MapTransactionMetadatumToTransactionMetadatum | void}
+   * @returns {MetadataMap | void}
    */
-  as_map_transaction_metadatum_to_transaction_metadatum(): MapTransactionMetadatumToTransactionMetadatum | void;
+  as_map(): MetadataMap | void;
 
   /**
-   * @returns {TransactionMetadatums | void}
+   * @returns {MetadataList | void}
    */
-  as_arr_transaction_metadatum(): TransactionMetadatums | void;
+  as_list(): MetadataList | void;
 
   /**
    * @returns {Int | void}
@@ -3402,43 +3448,6 @@ declare export class TransactionMetadatumLabels {
    * @param {BigNum} elem
    */
   add(elem: BigNum): void;
-}
-/**
- */
-declare export class TransactionMetadatums {
-  free(): void;
-
-  /**
-   * @returns {Uint8Array}
-   */
-  to_bytes(): Uint8Array;
-
-  /**
-   * @param {Uint8Array} bytes
-   * @returns {TransactionMetadatums}
-   */
-  static from_bytes(bytes: Uint8Array): TransactionMetadatums;
-
-  /**
-   * @returns {TransactionMetadatums}
-   */
-  static new(): TransactionMetadatums;
-
-  /**
-   * @returns {number}
-   */
-  len(): number;
-
-  /**
-   * @param {number} index
-   * @returns {TransactionMetadatum}
-   */
-  get(index: number): TransactionMetadatum;
-
-  /**
-   * @param {TransactionMetadatum} elem
-   */
-  add(elem: TransactionMetadatum): void;
 }
 /**
  */
