+++ rust/src/lib.rs
@@ -37,6 +37,7 @@ pub mod error;
 pub mod fees;
 pub mod impl_mockchain;
 pub mod legacy_address;
+pub mod metadata;
 pub mod serialization;
 pub mod tx_builder;
 pub mod typed_bytes;
@@ -46,6 +47,7 @@ pub mod utils;
 use address::*;
 use crypto::*;
 use error::*;
+use metadata::*;
 use utils::*;
 
 #[wasm_bindgen]
@@ -1171,249 +1173,6 @@ impl TransactionWitnessSet {
             scripts: None,
             bootstraps: None,
         }
-    }
-}
-
-#[wasm_bindgen]
-#[derive(Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
-pub struct MapTransactionMetadatumToTransactionMetadatum(
-    linked_hash_map::LinkedHashMap<TransactionMetadatum, TransactionMetadatum>,
-);
-
-to_from_bytes!(MapTransactionMetadatumToTransactionMetadatum);
-
-#[wasm_bindgen]
-impl MapTransactionMetadatumToTransactionMetadatum {
-    pub fn new() -> Self {
-        Self(linked_hash_map::LinkedHashMap::new())
-    }
-
-    pub fn len(&self) -> usize {
-        self.0.len()
-    }
-
-    pub fn insert(
-        &mut self,
-        key: &TransactionMetadatum,
-        value: &TransactionMetadatum,
-    ) -> Option<TransactionMetadatum> {
-        self.0.insert(key.clone(), value.clone())
-    }
-
-    pub fn get(&self, key: &TransactionMetadatum) -> Option<TransactionMetadatum> {
-        self.0.get(key).map(|v| v.clone())
-    }
-
-    pub fn keys(&self) -> TransactionMetadatums {
-        TransactionMetadatums(
-            self.0
-                .iter()
-                .map(|(k, _v)| k.clone())
-                .collect::<Vec<TransactionMetadatum>>(),
-        )
-    }
-}
-
-#[wasm_bindgen]
-#[derive(Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
-pub struct TransactionMetadatums(Vec<TransactionMetadatum>);
-
-to_from_bytes!(TransactionMetadatums);
-
-#[wasm_bindgen]
-impl TransactionMetadatums {
-    pub fn new() -> Self {
-        Self(Vec::new())
-    }
-
-    pub fn len(&self) -> usize {
-        self.0.len()
-    }
-
-    pub fn get(&self, index: usize) -> TransactionMetadatum {
-        self.0[index].clone()
-    }
-
-    pub fn add(&mut self, elem: &TransactionMetadatum) {
-        self.0.push(elem.clone());
-    }
-}
-
-#[wasm_bindgen]
-#[derive(Clone, Debug, Eq, Ord, PartialEq, PartialOrd)]
-pub enum TransactionMetadatumKind {
-    MapTransactionMetadatumToTransactionMetadatum,
-    ArrTransactionMetadatum,
-    Int,
-    Bytes,
-    Text,
-}
-
-#[derive(Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
-enum TransactionMetadatumEnum {
-    MapTransactionMetadatumToTransactionMetadatum(MapTransactionMetadatumToTransactionMetadatum),
-    ArrTransactionMetadatum(TransactionMetadatums),
-    Int(Int),
-    Bytes(Vec<u8>),
-    Text(String),
-}
-
-#[wasm_bindgen]
-#[derive(Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
-pub struct TransactionMetadatum(TransactionMetadatumEnum);
-
-to_from_bytes!(TransactionMetadatum);
-
-#[wasm_bindgen]
-impl TransactionMetadatum {
-    pub fn new_map_transaction_metadatum_to_transaction_metadatum(
-        map_transaction_metadatum_to_transaction_metadatum: &MapTransactionMetadatumToTransactionMetadatum,
-    ) -> Self {
-        Self(
-            TransactionMetadatumEnum::MapTransactionMetadatumToTransactionMetadatum(
-                map_transaction_metadatum_to_transaction_metadatum.clone(),
-            ),
-        )
-    }
-
-    pub fn new_arr_transaction_metadatum(
-        arr_transaction_metadatum: &TransactionMetadatums,
-    ) -> Self {
-        Self(TransactionMetadatumEnum::ArrTransactionMetadatum(
-            arr_transaction_metadatum.clone(),
-        ))
-    }
-
-    pub fn new_int(int: &Int) -> Self {
-        Self(TransactionMetadatumEnum::Int(int.clone()))
-    }
-
-    pub fn new_bytes(bytes: Vec<u8>) -> Self {
-        Self(TransactionMetadatumEnum::Bytes(bytes))
-    }
-
-    pub fn new_text(text: String) -> Self {
-        Self(TransactionMetadatumEnum::Text(text))
-    }
-
-    pub fn kind(&self) -> TransactionMetadatumKind {
-        match &self.0 {
-            TransactionMetadatumEnum::MapTransactionMetadatumToTransactionMetadatum(_) => {
-                TransactionMetadatumKind::MapTransactionMetadatumToTransactionMetadatum
-            }
-            TransactionMetadatumEnum::ArrTransactionMetadatum(_) => {
-                TransactionMetadatumKind::ArrTransactionMetadatum
-            }
-            TransactionMetadatumEnum::Int(_) => TransactionMetadatumKind::Int,
-            TransactionMetadatumEnum::Bytes(_) => TransactionMetadatumKind::Bytes,
-            TransactionMetadatumEnum::Text(_) => TransactionMetadatumKind::Text,
-        }
-    }
-
-    pub fn as_map_transaction_metadatum_to_transaction_metadatum(
-        &self,
-    ) -> Option<MapTransactionMetadatumToTransactionMetadatum> {
-        match &self.0 {
-            TransactionMetadatumEnum::MapTransactionMetadatumToTransactionMetadatum(x) => {
-                Some(x.clone())
-            }
-            _ => None,
-        }
-    }
-
-    pub fn as_arr_transaction_metadatum(&self) -> Option<TransactionMetadatums> {
-        match &self.0 {
-            TransactionMetadatumEnum::ArrTransactionMetadatum(x) => Some(x.clone()),
-            _ => None,
-        }
-    }
-
-    pub fn as_int(&self) -> Option<Int> {
-        match &self.0 {
-            TransactionMetadatumEnum::Int(x) => Some(x.clone()),
-            _ => None,
-        }
-    }
-
-    pub fn as_bytes(&self) -> Option<Vec<u8>> {
-        match &self.0 {
-            TransactionMetadatumEnum::Bytes(x) => Some(x.clone()),
-            _ => None,
-        }
-    }
-
-    pub fn as_text(&self) -> Option<String> {
-        match &self.0 {
-            TransactionMetadatumEnum::Text(x) => Some(x.clone()),
-            _ => None,
-        }
-    }
-}
-
-type TransactionMetadatumLabel = BigNum;
-
-#[wasm_bindgen]
-#[derive(Clone, Debug, Eq, Ord, PartialEq, PartialOrd)]
-pub struct TransactionMetadatumLabels(Vec<TransactionMetadatumLabel>);
-
-to_from_bytes!(TransactionMetadatumLabels);
-
-#[wasm_bindgen]
-impl TransactionMetadatumLabels {
-    pub fn new() -> Self {
-        Self(Vec::new())
-    }
-
-    pub fn len(&self) -> usize {
-        self.0.len()
-    }
-
-    pub fn get(&self, index: usize) -> TransactionMetadatumLabel {
-        self.0[index].clone()
-    }
-
-    pub fn add(&mut self, elem: &TransactionMetadatumLabel) {
-        self.0.push(elem.clone());
-    }
-}
-
-#[wasm_bindgen]
-#[derive(Clone, Debug, Eq, Ord, PartialEq, PartialOrd)]
-pub struct TransactionMetadata(
-    linked_hash_map::LinkedHashMap<TransactionMetadatumLabel, TransactionMetadatum>,
-);
-
-to_from_bytes!(TransactionMetadata);
-
-#[wasm_bindgen]
-impl TransactionMetadata {
-    pub fn new() -> Self {
-        Self(linked_hash_map::LinkedHashMap::new())
-    }
-
-    pub fn len(&self) -> usize {
-        self.0.len()
-    }
-
-    pub fn insert(
-        &mut self,
-        key: &TransactionMetadatumLabel,
-        value: &TransactionMetadatum,
-    ) -> Option<TransactionMetadatum> {
-        self.0.insert(key.clone(), value.clone())
-    }
-
-    pub fn get(&self, key: &TransactionMetadatumLabel) -> Option<TransactionMetadatum> {
-        self.0.get(key).map(|v| v.clone())
-    }
-
-    pub fn keys(&self) -> TransactionMetadatumLabels {
-        TransactionMetadatumLabels(
-            self.0
-                .iter()
-                .map(|(k, _v)| k.clone())
-                .collect::<Vec<TransactionMetadatumLabel>>(),
-        )
     }
 }
