+++ rust/src/serialization.rs
@@ -1682,204 +1682,6 @@ impl DeserializeEmbeddedGroup for TransactionWitnessSet {
     }
 }
 
-impl cbor_event::se::Serialize for MapTransactionMetadatumToTransactionMetadatum {
-    fn serialize<'se, W: Write>(&self, serializer: &'se mut Serializer<W>) -> cbor_event::Result<&'se mut Serializer<W>> {
-        serializer.write_map(cbor_event::Len::Len(self.0.len() as u64))?;
-        for (key, value) in &self.0 {
-            key.serialize(serializer)?;
-            value.serialize(serializer)?;
-        }
-        Ok(serializer)
-    }
-}
-
-impl Deserialize for MapTransactionMetadatumToTransactionMetadatum {
-    fn deserialize<R: BufRead + Seek>(raw: &mut Deserializer<R>) -> Result<Self, DeserializeError> {
-        let mut table = linked_hash_map::LinkedHashMap::new();
-        (|| -> Result<_, DeserializeError> {
-            let len = raw.map()?;
-            while match len { cbor_event::Len::Len(n) => table.len() < n as usize, cbor_event::Len::Indefinite => true, } {
-                if raw.cbor_type()? == CBORType::Special {
-                    assert_eq!(raw.special()?, CBORSpecial::Break);
-                    break;
-                }
-                let key = TransactionMetadatum::deserialize(raw)?;
-                let value = TransactionMetadatum::deserialize(raw)?;
-                if table.insert(key.clone(), value).is_some() {
-                    return Err(DeserializeFailure::DuplicateKey(Key::Str(String::from("some complicated/unsupported type"))).into());
-                }
-            }
-            Ok(())
-        })().map_err(|e| e.annotate("MapTransactionMetadatumToTransactionMetadatum"))?;
-        Ok(Self(table))
-    }
-}
-
-impl cbor_event::se::Serialize for TransactionMetadatums {
-    fn serialize<'se, W: Write>(&self, serializer: &'se mut Serializer<W>) -> cbor_event::Result<&'se mut Serializer<W>> {
-        serializer.write_array(cbor_event::Len::Len(self.0.len() as u64))?;
-        for element in &self.0 {
-            element.serialize(serializer)?;
-        }
-        Ok(serializer)
-    }
-}
-
-impl Deserialize for TransactionMetadatums {
-    fn deserialize<R: BufRead + Seek>(raw: &mut Deserializer<R>) -> Result<Self, DeserializeError> {
-        let mut arr = Vec::new();
-        (|| -> Result<_, DeserializeError> {
-            let len = raw.array()?;
-            while match len { cbor_event::Len::Len(n) => arr.len() < n as usize, cbor_event::Len::Indefinite => true, } {
-                if raw.cbor_type()? == CBORType::Special {
-                    assert_eq!(raw.special()?, CBORSpecial::Break);
-                    break;
-                }
-                arr.push(TransactionMetadatum::deserialize(raw)?);
-            }
-            Ok(())
-        })().map_err(|e| e.annotate("TransactionMetadatums"))?;
-        Ok(Self(arr))
-    }
-}
-
-impl cbor_event::se::Serialize for TransactionMetadatumEnum {
-    fn serialize<'se, W: Write>(&self, serializer: &'se mut Serializer<W>) -> cbor_event::Result<&'se mut Serializer<W>> {
-        match self {
-            TransactionMetadatumEnum::MapTransactionMetadatumToTransactionMetadatum(x) => {
-                x.serialize(serializer)
-            },
-            TransactionMetadatumEnum::ArrTransactionMetadatum(x) => {
-                x.serialize(serializer)
-            },
-            TransactionMetadatumEnum::Int(x) => {
-                x.serialize(serializer)
-            },
-            TransactionMetadatumEnum::Bytes(x) => {
-                serializer.write_bytes(&x)
-            },
-            TransactionMetadatumEnum::Text(x) => {
-                serializer.write_text(&x)
-            },
-        }
-    }
-}
-
-impl Deserialize for TransactionMetadatumEnum {
-    fn deserialize<R: BufRead + Seek>(raw: &mut Deserializer<R>) -> Result<Self, DeserializeError> {
-        let initial_position = raw.as_mut_ref().seek(SeekFrom::Current(0)).unwrap();
-        match (|raw: &mut Deserializer<_>| -> Result<_, DeserializeError> {
-            Ok(MapTransactionMetadatumToTransactionMetadatum::deserialize(raw)?)
-        })(raw)
-        {
-            Ok(variant) => return Ok(TransactionMetadatumEnum::MapTransactionMetadatumToTransactionMetadatum(variant)),
-            Err(_) => raw.as_mut_ref().seek(SeekFrom::Start(initial_position)).unwrap(),
-        };
-        match (|raw: &mut Deserializer<_>| -> Result<_, DeserializeError> {
-            Ok(TransactionMetadatums::deserialize(raw)?)
-        })(raw)
-        {
-            Ok(variant) => return Ok(TransactionMetadatumEnum::ArrTransactionMetadatum(variant)),
-            Err(_) => raw.as_mut_ref().seek(SeekFrom::Start(initial_position)).unwrap(),
-        };
-        match (|raw: &mut Deserializer<_>| -> Result<_, DeserializeError> {
-            Ok(Int::deserialize(raw)?)
-        })(raw)
-        {
-            Ok(variant) => return Ok(TransactionMetadatumEnum::Int(variant)),
-            Err(_) => raw.as_mut_ref().seek(SeekFrom::Start(initial_position)).unwrap(),
-        };
-        match (|raw: &mut Deserializer<_>| -> Result<_, DeserializeError> {
-            Ok(raw.bytes()?)
-        })(raw)
-        {
-            Ok(variant) => return Ok(TransactionMetadatumEnum::Bytes(variant)),
-            Err(_) => raw.as_mut_ref().seek(SeekFrom::Start(initial_position)).unwrap(),
-        };
-        match (|raw: &mut Deserializer<_>| -> Result<_, DeserializeError> {
-            Ok(String::deserialize(raw)?)
-        })(raw)
-        {
-            Ok(variant) => return Ok(TransactionMetadatumEnum::Text(variant)),
-            Err(_) => raw.as_mut_ref().seek(SeekFrom::Start(initial_position)).unwrap(),
-        };
-        Err(DeserializeError::new("TransactionMetadatumEnum", DeserializeFailure::NoVariantMatched.into()))
-    }
-}
-
-impl cbor_event::se::Serialize for TransactionMetadatum {
-    fn serialize<'se, W: Write>(&self, serializer: &'se mut Serializer<W>) -> cbor_event::Result<&'se mut Serializer<W>> {
-        self.0.serialize(serializer)
-    }
-}
-
-impl Deserialize for TransactionMetadatum {
-    fn deserialize<R: BufRead + Seek>(raw: &mut Deserializer<R>) -> Result<Self, DeserializeError> {
-        Ok(Self(TransactionMetadatumEnum::deserialize(raw)?))
-    }
-}
-
-impl cbor_event::se::Serialize for TransactionMetadatumLabels {
-    fn serialize<'se, W: Write>(&self, serializer: &'se mut Serializer<W>) -> cbor_event::Result<&'se mut Serializer<W>> {
-        serializer.write_array(cbor_event::Len::Len(self.0.len() as u64))?;
-        for element in &self.0 {
-            element.serialize(serializer)?;
-        }
-        Ok(serializer)
-    }
-}
-
-impl Deserialize for TransactionMetadatumLabels {
-    fn deserialize<R: BufRead + Seek>(raw: &mut Deserializer<R>) -> Result<Self, DeserializeError> {
-        let mut arr = Vec::new();
-        (|| -> Result<_, DeserializeError> {
-            let len = raw.array()?;
-            while match len { cbor_event::Len::Len(n) => arr.len() < n as usize, cbor_event::Len::Indefinite => true, } {
-                if raw.cbor_type()? == CBORType::Special {
-                    assert_eq!(raw.special()?, CBORSpecial::Break);
-                    break;
-                }
-                arr.push(TransactionMetadatumLabel::deserialize(raw)?);
-            }
-            Ok(())
-        })().map_err(|e| e.annotate("TransactionMetadatumLabels"))?;
-        Ok(Self(arr))
-    }
-}
-
-impl cbor_event::se::Serialize for TransactionMetadata {
-    fn serialize<'se, W: Write>(&self, serializer: &'se mut Serializer<W>) -> cbor_event::Result<&'se mut Serializer<W>> {
-        serializer.write_map(cbor_event::Len::Len(self.0.len() as u64))?;
-        for (key, value) in &self.0 {
-            key.serialize(serializer)?;
-            value.serialize(serializer)?;
-        }
-        Ok(serializer)
-    }
-}
-
-impl Deserialize for TransactionMetadata {
-    fn deserialize<R: BufRead + Seek>(raw: &mut Deserializer<R>) -> Result<Self, DeserializeError> {
-        let mut table = linked_hash_map::LinkedHashMap::new();
-        (|| -> Result<_, DeserializeError> {
-            let len = raw.map()?;
-            while match len { cbor_event::Len::Len(n) => table.len() < n as usize, cbor_event::Len::Indefinite => true, } {
-                if raw.cbor_type()? == CBORType::Special {
-                    assert_eq!(raw.special()?, CBORSpecial::Break);
-                    break;
-                }
-                let key = TransactionMetadatumLabel::deserialize(raw)?;
-                let value = TransactionMetadatum::deserialize(raw)?;
-                if table.insert(key.clone(), value).is_some() {
-                    return Err(DeserializeFailure::DuplicateKey(Key::Str(String::from("some complicated/unsupported type"))).into());
-                }
-            }
-            Ok(())
-        })().map_err(|e| e.annotate("TransactionMetadata"))?;
-        Ok(Self(table))
-    }
-}
-
 impl cbor_event::se::Serialize for MsigPubkey {
     fn serialize<'se, W: Write>(&self, serializer: &'se mut Serializer<W>) -> cbor_event::Result<&'se mut Serializer<W>> {
         serializer.write_array(cbor_event::Len::Len(2))?;
